Merge upstream/master

+++ cpukit/score/include/rtems/score/heap.h
@@ -360,17 +360,17 @@ typedef struct {
   /**
    * @brief Number of blocks of this type.
    */
-  uint32_t number;
+  uintptr_t number;
 
   /**
    * @brief Largest block of this type.
    */
-  uint32_t largest;
+  uintptr_t largest;
 
   /**
    * @brief Total size of the blocks of this type.
    */
-  uint32_t total;
+  uintptr_t total;
 } Heap_Information;
 
 /**
@@ -413,11 +413,12 @@ typedef uintptr_t (*Heap_Initialization_or_extend_handler)(
  * @brief Extends the memory available for the heap @a heap using the memory
  * area starting at @a area_begin of size @a area_size bytes.
  *
- * There are no alignment requirements.  The memory area must be big enough to
- * contain some maintainance blocks.  It must not overlap parts of the current
- * heap areas.  Disconnected subordinate heap areas will lead to used blocks
- * which cover the gaps.  Extending with an inappropriate memory area will
- * corrupt the heap.
+ * There are no alignment requirements for the memory area.  The memory area
+ * must be big enough to contain some maintenance blocks.  It must not overlap
+ * parts of the current heap memory areas.  Disconnected memory areas added to
+ * the heap will lead to used blocks which cover the gaps.  Extending with an
+ * inappropriate memory area will corrupt the heap resulting in undefined
+ * behaviour.
  *
  * The unused fourth parameter is provided to have the same signature as
  * _Heap_Initialize().
